Feature5: Updated base.html to display different menu for logged in users, and added dropdown with profile options. Adjusted login and register views to properly handle user authentication.
