+++ donations/static/css/style.css
@@ -1946,3 +1946,29 @@ footer .bottom-line--icons a img {
 .error-message li {
     margin-bottom: 5px;
 }
+
+/* static/css/style.css */
+/* Add this to your existing CSS file */
+.dropdown-content {
+    display: none;
+    position: absolute;
+    background-color: #f9f9f9;
+    min-width: 160px;
+    box-shadow: 0px 8px 16px 0px rgba(0,0,0,0.2);
+    z-index: 1;
+}
+
+.dropdown-content a {
+    color: black;
+    padding: 12px 16px;
+    text-decoration: none;
+    display: block;
+}
+
+.dropdown-content a:hover {
+    background-color: #f1f1f1
+}
+
+.dropdown:hover .dropdown-content {
+    display: block;
+}
